complete rework of the bot
