--- .env
@@ -1,1 +1,0 @@
-DISCORD_TOKEN=
